Fixed sneaky bug that was left from the old meteor-enc implementation: subscription.ready() was sometimes called before the data was decrypted. The meteor-enc code that was trying to prevent this was assuming that subscription names equal collection names, but this is not true for search, and in general.
A fix would be to determine the name of a collection for a subscription to call the decrypt callbacks of the collection, but a subscription could be to multiple collections (e.g., if one uses self.added), and Meteor does nto record such information.

The right solution would be to have decrypt callbacks (_decrypt_cb) per subscription (instead of collection) and make sure that subscription.ready() is only called when those callbacks are ready. However, when decrypting a document in a collection, there is no way to tell what subscription the document is for: the wire protocol only seems to include collection name for data messages.

For now, I just did a quick fix that works for subscriptions to only one collection. I can determine the name of the collection from the subscription record.