Better fix for issue with decrypt callbacks brought about by the lack of information about collections in subscription _process_ready. This issue meant that we might call ready before decrypting all messages for subscriptions with names different than the collections names.
Turns out that in Meteor, a DDP message does not belong to a subscription from the client's standpoint, which was the confusing point. Only the server knows this mapping. The server sends messages of the form to the client:

start_subscription subID
add collection content
start_subscription subID2
add collection content    <-- this could belong to both subscriptions
end_subscription subID
end_subscription subID2

Therefore, having a decrypt_cb per collection does not make sense (or per subscription) because when the subscription is ready we don't know which collection decrypt_cb to use.

The fix is to just have one global mylar_decrypt_cb list and any subscription ready callback will be called only after the decryption finished for all messages that started decryption before the end_subscription. To avoid deadlock, we need to make sure there will be no circular dependencies, so for subscriptions that dec_msg relies on, we just run the ready callback because they don't do decryptions. We need to have an up-to-date list of such subscriptions which means we need to update this list on future changes, which is not ideal. The right fix is to change the wire protocol and have the server tell us at the end of a subscription what were the collections affected by the subscription.

We now make even less changes to Meteor; no need for getCollection, _getCollection changes in collection/livedata.