Rewrote outbound intercept code. We were interposing too early on outgoing data.
A small issue in Medtest (the required questions showed incomplete for a while after they were completed) pointed to this bigger issue: we were encrypting data before it was being placed in the minicollection and sent to the server. We should instead encrypt just after it is placed in the minicollection because:

1. we don't want encrypted data in the minicollection

2. we want updates in the mini collection to be immediately available/synchronous -- as opposed to waiting until encryption completes (which is asynchronous due to principal lookup)
  --> this was the issue with MedTest: after inserting answers locally, it was fetching these answers from the local collection to see if they were complete; but the asynchronous encryption did not finish by then so it would not see the new result

Meteor's code wasn't clean here: they were doing the local and remote insert in the same function livedata_connection.js:apply and a bit intermingled. I changed this function to separate the two, do the local update, and only then call the intercept code for encryption (before sending data to the server). I plan to clean this code even more -- onto the TODO list.

MedTest issue fixed.